pyproject.toml
[project]
name = "ultrasync-mcp"
version = "0.1.0"
description = "Semantic code indexing and memory for AI coding agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "David Gidwani", email = "david.gidwani@gmail.com" }]
keywords = ["ai", "coding-agent", "embeddings", "mcp", "semantic-search"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "hyperscan>=0.8.0",
  "lmdb>=1.5.1",
  "mcp>=1.2.0",
  "msgpack>=1.1.0",
  "numpy>=1.24",
  "sentence-transformers>=3",
  "structlog>=24.0",
  "tyro>=0.9",
  "ultrasync-index",
  "urllib3>=2.6.0",
]

[project.urls]
Homepage = "https://github.com/davidgidwani/ultrasync"
Repository = "https://github.com/davidgidwani/ultrasync"

[project.scripts]
ultrasync = "ultrasync_mcp._entry:main"

[project.optional-dependencies]
benchmark = [
  "pyyaml>=6.0",  # for loading test cases
]
cli = [
  "rich>=13.0",
  "textual>=0.50",
]
infinity = [
  "httpx>=0.27",  # for InfinityAPIProvider (remote server)
  # infinity-emb excluded - pins numpy<2 which breaks Python 3.14t
  # install manually with: uv pip install infinity-emb --no-deps
]
lexical = [
  "tantivy>=0.22",  # BM25 lexical search via Rust bindings
]
secrets = [
  "detect-secrets>=1.4.0",  # enterprise secret detection by Yelp
]
sync = [
  "python-socketio[asyncio_client]>=5.11.0",  # async socket.io for sync server
]

[dependency-groups]
dev = [
  "ipython>=8.37.0",
  "maturin>=1.5",
  "pytest>=9.0.1",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.14.8",
  "tombi>=0.7.7",
]
docs = [
  "zensical>=0.0.11",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = [
  "B101",  # assert used for internal checks
  "B105",  # false positives on env var names (ULTRASYNC_*_TOKEN etc)
  "B110",  # try/except/pass is intentional for optional features
  "B112",  # try/except/continue is intentional for batch processing
  "B404",  # subprocess import is required for git operations
  "B603",  # subprocess calls are safe (hardcoded git commands)
  "B606",  # os.execv with known paths is intentional
  "B607",  # partial paths to git are acceptable
]

[tool.pytest]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "TID252"]

[tool.ruff.lint.isort]
known-first-party = ["ultrasync_mcp", "ultrasync_index"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]  # allow imports after warnings filter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.uv.sources]
ultrasync-index = { path = "crates/ultrasync_index", editable = true }
